+++ spec/castkit/core/attribute_types_spec.rb
@@ -3,84 +3,89 @@
 require "spec_helper"
 
 RSpec.describe Castkit::Core::AttributeTypes do
-  let(:klass) do
-    Class.new do
-      extend Castkit::Core::AttributeTypes
-
-      def self.attribute(field, type, **options)
-        @attributes ||= []
-        @attributes << { field: field, type: type, options: options }
-      end
-
-      class << self
-        attr_reader :attributes
-      end
+  it "extends including class with type helpers and initializes config" do
+    klass = Class.new do
+      include Castkit::Core::AttributeTypes
+      extend Castkit::Core::Attributes
     end
+
+    klass.string :name
+    instance = klass.new
+    instance.name = "hi"
+
+    expect(instance.name).to eq("hi")
   end
 
-  describe "type methods" do
-    it "registers a string attribute" do
-      klass.string :name, required: false
-      expect(klass.attributes.last).to eq({ field: :name, type: :string, options: { required: false } })
+  it "supports dynamic DSL definition for registered types" do
+    klass = Class.new do
+      extend Castkit::Core::Attributes
+      include Castkit::Core::AttributeTypes
     end
 
-    it "registers an integer attribute" do
-      klass.integer :age, min: 0
-      expect(klass.attributes.last).to eq({ field: :age, type: :integer, options: { min: 0 } })
+    Castkit::Core::AttributeTypes.define_type_dsl(:custom_type)
+    allow(klass).to receive(:attribute)
+
+    klass.custom_type :field, foo: :bar
+    expect(klass).to have_received(:attribute).with(:field, :custom_type, foo: :bar)
+  end
+
+  it "falls back to Object#hash when called without arguments" do
+    klass = Class.new do
+      include Castkit::Core::AttributeTypes
+
+      def attribute(*); end
     end
 
-    it "registers a boolean attribute" do
-      klass.boolean :admin
-      expect(klass.attributes.last).to eq({ field: :admin, type: :boolean, options: {} })
+    expect { klass.hash }.not_to raise_error
+  end
+
+  it "raises when dataobject type is not a subclass of Castkit::DataObject" do
+    klass = Class.new do
+      include Castkit::Core::AttributeTypes
+      extend Castkit::Core::Attributes
     end
 
-    it "registers a float attribute" do
-      klass.float :price
-      expect(klass.attributes.last).to eq({ field: :price, type: :float, options: {} })
+    expect do
+      klass.dataobject(:bad, String)
+    end.to raise_error(Castkit::AttributeError, /must extend/)
+  end
+
+  it "defines helper methods for all built-in types" do
+    recorded = []
+    klass = Class.new do
+      extend Castkit::Core::Attributes
+      include Castkit::Core::AttributeTypes
+
+      define_singleton_method(:attribute) { |field, type, **| recorded << [field, type] }
     end
 
-    it "registers a date attribute" do
-      klass.date :dob
-      expect(klass.attributes.last).to eq({ field: :dob, type: :date, options: {} })
+    klass.integer :a
+    klass.boolean :b
+    klass.float :c
+    klass.date :d
+    klass.datetime :e
+    klass.array :f, of: :string
+    klass.hash :g
+    dataobject_type = Class.new(Castkit::DataObject)
+    klass.dataobject(:h, dataobject_type)
+    klass.unwrapped(:h, dataobject_type)
+    klass.hash
+
+    expect(recorded).to include(
+      %i[a integer],
+      %i[b boolean],
+      %i[c float],
+      %i[d date],
+      %i[e datetime],
+      %i[f array],
+      %i[g hash]
+    )
+
+    entries = recorded.each_with_object(Hash.new { |hash, key| hash[key] = [] }) do |(field, type), memo|
+      memo[field] << type
     end
 
-    it "registers a datetime attribute" do
-      klass.datetime :created_at
-      expect(klass.attributes.last).to eq({ field: :created_at, type: :datetime, options: {} })
-    end
-
-    it "registers an array attribute" do
-      klass.array :tags
-      expect(klass.attributes.last).to eq({ field: :tags, type: :array, options: {} })
-    end
-
-    it "registers a hash attribute" do
-      klass.hash :metadata
-      expect(klass.attributes.last).to eq({ field: :metadata, type: :hash, options: {} })
-    end
-
-    it "registers a dataobject attribute" do
-      dto_class = Class.new(Castkit::DataObject)
-      klass.dataobject :profile, dto_class, required: false
-      expect(klass.attributes.last).to eq({ field: :profile, type: dto_class, options: { required: false } })
-    end
-
-    it "raises if dataobject type is not a subclass of Castkit::DataObject" do
-      expect do
-        klass.dataobject :bad, String
-      end.to raise_error(Castkit::AttributeError, /must extend from Castkit::DataObject/)
-    end
-
-    it "registers an unwrapped dataobject" do
-      dto_class = Class.new(Castkit::DataObject)
-      klass.unwrapped :location, dto_class
-      expect(klass.attributes.last).to eq({ field: :location, type: dto_class, options: { unwrapped: true } })
-    end
-
-    it "aliases dto to dataobject" do
-      dto_class = Class.new(Castkit::DataObject)
-      klass.dto :thing, dto_class
-      expect(klass.attributes.last).to eq({ field: :thing, type: dto_class, options: {} })
-    end
+    expect(entries[:h].size).to eq(2)
+    expect(entries[:h]).to all(eq(dataobject_type))
   end
 end